+++ src/components/UserSettings/Requests/index.jsx
@@ -1,44 +1,46 @@
-import { func, node } from "prop-types";
-import { createContext, useCallback, useContext, useMemo, useState } from "react";
+import { useCallback } from "react";
+import { Button, Dialog } from "@blueprintjs/core";
 
-import * as customPropTypes from "../../../propTypes";
+import { RowCol, SettingsManager } from "Components/UserSettings";
+import DataRequest from "./DataRequest";
+import RequestsEditor from "./RequestsEditor";
+
+import { useBool } from "../../../hooks";
+import { CustomClasses } from "../../../constants";
 
 
-const RequestsSettings = createContext({});
+const { Provider: RequestsProvider, useSettings: useRequestsSettings } = new SettingsManager({
+	afterUpdate: (prevState, update) => {
+		window?.zoteroRoam?.updateLibraries?.(update.libraries);
+	}
+});
 
-const RequestsProvider = ({ children, init, updater }) => {
-	const [requests, _setRequests] = useState(init);
+function RequestsWidget() {
+	const [
+		{
+			/*apiKeys,*/
+			dataRequests,
+			/*libraries*/
+		},
+		setOpts
+	] = useRequestsSettings();
 
-	const setRequests = useCallback((updateFn) => {
-		_setRequests((prevState) => {
-			const update = updateFn(prevState);
-			updater(update);
-			window?.zoteroRoam?.updateLibraries?.(update.libraries);
-			return update;
-		});
-	}, [updater]);
+	const [isDialogOpen, { on: openDialog, off: closeDialog }] = useBool(false);
 
-	const contextValue = useMemo(() => [requests, setRequests], [requests, setRequests]);
+	const updateRequests = useCallback((val) => setOpts(_prevState => val), [setOpts]);
 
-	return (
-		<RequestsSettings.Provider value={contextValue}>
-			{children}
-		</RequestsSettings.Provider>
-	);
-};
-RequestsProvider.propTypes = {
-	children: node,
-	init: customPropTypes.requestsType,
-	updater: func
-};
-
-const useRequestsSettings = () => {
-	const context = useContext(RequestsSettings);
-
-	return context;
-};
+	return <>
+		<RowCol title="Data Requests" description="Choose which items to sync from Zotero" rightElement={<Button className={CustomClasses.TEXT_SMALL} icon={dataRequests.length == 0 ? "arrow-right" : null} intent="warning" minimal={true} onClick={openDialog} text={dataRequests.length == 0 ? "Add request" : "Edit"} />}>
+			{dataRequests.map((req) => <DataRequest key={req.name} request={req} />)}
+		</RowCol>
+		<Dialog className="zr-data-requests-dialog" isOpen={isDialogOpen} onClose={closeDialog}>
+			<RequestsEditor closeDialog={closeDialog} dataRequests={dataRequests} updateRequests={updateRequests} />
+		</Dialog>
+	</>;
+}
 
 export {
 	RequestsProvider,
+	RequestsWidget,
 	useRequestsSettings
 };
